% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsnet_cols.R
\name{pedsnet_cols}
\alias{pedsnet_cols}
\title{Function to extract pedsnet colors as hex codes}
\usage{
pedsnet_cols(...)
}
\arguments{
\item{...}{Character names of pedsnet_colors}
}
\description{
Function to extract pedsnet colors as hex codes
}
