% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsnet_pal.R
\name{pedsnet_pal}
\alias{pedsnet_pal}
\title{Return function to interpolate a pedsnet color palette}
\usage{
pedsnet_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pedsnet_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a pedsnet color palette
}
